Large rename of files to make filenames consistent
